bites/requirements.txt
django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.8
channels==4.1.0
channels-redis==4.2.0
google-api-python-client==2.149.0
yt-dlp==2024.8.6
python-dotenv==1.0.1
gunicorn==23.0.0
supabase==2.8.1
django-celery-results==2.5.1
isodate==0.6.1
transformers>=4.36.0
torch>=2.0.0

